Largely refactored Shape architecture - made more simple and more convinient

+++ src/shared/Point2D.h
@@ -1,9 +1,9 @@
 #pragma once
 struct Point2D {
-	long int x;
-	long int y;
+	int x;
+	int y;
 
-	long int distanceToInSquare(Point2D other)
+	int distanceToInSquare(Point2D other)
 	{
 		return ((x - other.x) * (x - other.x) + (y - other.y) * (y - other.y));
 	}
